Added base query builder converter
- Introduced ConverterInterface
- Type hint interface not base class
- Removed "Builder" prefix from query builder converters
- Removed unused use statements
- Removed swap file
