+++ lib/Doctrine/ODM/PHPCR/Id/AssignedIdGenerator.php
@@ -20,7 +20,6 @@
 namespace Doctrine\ODM\PHPCR\Id;
 
 use Doctrine\ODM\PHPCR\DocumentManagerInterface;
-use Doctrine\ODM\PHPCR\PHPCRException;
 use Doctrine\ODM\PHPCR\Mapping\ClassMetadata;
 
 /**
