+++ lib/Doctrine/ODM/PHPCR/Mapping/Annotations/Reference.php
@@ -19,7 +19,6 @@
 
 namespace Doctrine\ODM\PHPCR\Mapping\Annotations;
 
-use Doctrine\Common\Annotations\Annotation;
 
 /**
  * base class for the reference types
