+++ lib/Doctrine/ODM/PHPCR/Query/Builder/OperandStaticFactory.php
@@ -2,7 +2,6 @@
 
 namespace Doctrine\ODM\PHPCR\Query\Builder;
 
-use PHPCR\Query\QOM\QueryObjectModelConstantsInterface as QOMConstants;
 
 /**
  * Factory node for static operands.
