+++ lib/Doctrine/ODM/PHPCR/Query/Builder/OrderByAdd.php
@@ -2,7 +2,6 @@
 
 namespace Doctrine\ODM\PHPCR\Query\Builder;
 
-use PHPCR\Query\QOM\QueryObjectModelConstantsInterface as QOMConstants;
 
 /**
  * Node for adding an additional order by.
