+++ lib/Doctrine/ODM/PHPCR/ReferenceManyCollection.php
@@ -22,7 +22,6 @@ namespace Doctrine\ODM\PHPCR;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Util\ClassUtils;
-use PHPCR\NodeInterface;
 use Doctrine\ODM\PHPCR\Exception\InvalidArgumentException;
 
 /**
