+++ lib/Doctrine/ODM/PHPCR/Translation/TranslationStrategy/AttributeTranslationStrategy.php
@@ -22,7 +22,6 @@ namespace Doctrine\ODM\PHPCR\Translation\TranslationStrategy;
 use Doctrine\ODM\PHPCR\Mapping\ClassMetadata,
     PHPCR\NodeInterface;
 
-use Doctrine\ODM\PHPCR\Query\Builder\SourceFactory;
 use Doctrine\ODM\PHPCR\Translation\Translation;
 use PHPCR\Query\QOM\ConstraintInterface;
 use PHPCR\Query\QOM\QueryObjectModelFactoryInterface;
